+++ packages/cli/src/index.ts
@@ -14,6 +14,7 @@ import {
   serveSite,
   generateCodeForProject,
   findDuplicateSegmentsInProject,
+  findUsageInProject,
   BuildCLIOptions,
   GenerateCodeCLIOptions,
   TestProjectOptions,
@@ -22,6 +23,9 @@ import {
   Dependencies,
   Datasource,
   openGui,
+  benchmarkFeature,
+  showProjectConfig,
+  evaluateFeature,
 } from "@featurevisor/core";
 
 process.on("unhandledRejection", (reason) => {
@@ -231,7 +235,138 @@ async function main() {
         }
       },
     })
-    .example("$0 generate-code", "generate code from YAMLs")
+    .example("$0 find-duplicate-segments", "list segments with same conditions")
+
+    /**
+     * Find usage
+     */
+    .command({
+      command: "find-usage",
+      handler: async function (options) {
+        const deps = await getDependencies(options);
+
+        try {
+          await findUsageInProject(deps, {
+            segment: options.segment,
+            attribute: options.attribute,
+
+            unusedSegments: options.unusedSegments,
+            unusedAttributes: options.unusedAttributes,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 find-usage", "find usage of segments and attributes")
+    .example("$0 find-usage --segment=my_segment", "find usage of segment")
+    .example("$0 find-usage --attribute=my_attribute", "find usage of attribute")
+    .example("$0 find-usage --unusedSegments", "find unused segments")
+    .example("$0 find-usage --unusedAttributes", "find unused attributes")
+
+    /**
+     * Benchmark features
+     */
+    .command({
+      command: "benchmark",
+      handler: async function (options) {
+        if (!options.environment) {
+          console.error("Please specify an environment with --environment flag.");
+          process.exit(1);
+        }
+
+        if (!options.feature) {
+          console.error("Please specify a feature with --feature flag.");
+          process.exit(1);
+        }
+
+        const deps = await getDependencies(options);
+
+        try {
+          await benchmarkFeature(deps, {
+            environment: options.environment,
+            feature: options.feature,
+            n: parseInt(options.n, 10) || 1,
+            context: options.context ? JSON.parse(options.context) : {},
+            variation: options.variation || undefined,
+            variable: options.variable || undefined,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 benchmark", "benchmark feature evaluations")
+    .example(
+      "$0 benchmark --environment=production --feature=my_feature --context='{}' -n=100",
+      "benchmark feature flag evaluation",
+    )
+    .example(
+      "$0 benchmark --environment=production --feature=my_feature --context='{}' --variation -n=100",
+      "benchmark feature variation evaluation",
+    )
+    .example(
+      "$0 benchmark --environment=production --feature=my_feature --context='{}' --variable=my_variable_key -n=100",
+      "benchmark feature variable evaluation",
+    )
+
+    /**
+     * Show config
+     */
+    .command({
+      command: "config",
+      handler: async function (options) {
+        const projectConfig = requireAndGetProjectConfig(rootDirectoryPath);
+
+        showProjectConfig(projectConfig, {
+          print: options.print,
+          pretty: options.pretty,
+        });
+      },
+    })
+    .example("$0 config", "show project configuration")
+    .example("$0 config --print", "print project configuration as JSON")
+    .example("$0 config --print --pretty", "print project configuration as prettified JSON")
+
+    /**
+     * Evaluate
+     */
+    .command({
+      command: "evaluate",
+      handler: async function (options) {
+        if (!options.environment) {
+          console.error("Please specify an environment with --environment flag.");
+          process.exit(1);
+        }
+
+        if (!options.feature) {
+          console.error("Please specify a feature with --feature flag.");
+          process.exit(1);
+        }
+
+        const deps = await getDependencies(options);
+
+        try {
+          await evaluateFeature(deps, {
+            environment: options.environment,
+            feature: options.feature,
+            context: options.context ? JSON.parse(options.context) : {},
+            print: options.print,
+            pretty: options.pretty,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 evaluate", "evaluate a feature along with its variation and variables")
+    .example(
+      "$0 evaluate --environment=production --feature=my_feature --context='{}'",
+      "evaluate a feature against provided context",
+    )
 
     /**
      * GUI